+ Game: onUpdate renombrado a onStep. + GfxEngine: Actualizada especificación en .h, eliminado un puntero en setScreenBackgroundColor(...), corregido funcionamiento de setGameScreenScale, setGameScreenSize, setScreenSize y setScreenBackgroundColor; arreglados algunos warnings de autocast a float y añadido método renderRectangle(...) muy útil para depuración. + Image: Adaptado orden de parámetros de constructora que no carga para que sea consistente con el resto del sistema (gfxEngine ya no es el 1º parámetro) + Canvas modificado acorde al cambio anterior. + TileMap modificado acorde al cambio anterior y al nuevo renderExt de GfxEngine. + Actualizados objetos de testeo con nuevos métodos y funcionalidades. Por ejemplo todos pintan sus máscaras para depuración. + Añadidas dll del sonido necesarias. Están en la carpeta "interpreter/extlibs". Ponedlas junto al exe cuando compiléis.
Nota! Que nadie suba ejecutables ni .obj ni nada de depuración al SVN aún, eh!

+++ interpreter/source/testeo/objBad.h
@@ -28,7 +28,7 @@ class objBad : public Entity
 	void onStep()
 	{
 		counter++;
-		graphic->setRotation(sin(counter*3.14/180)*25);
+		graphic->setRotation((float) sin(counter*3.14/180)*25);
 	}
 };
